Improved test coverage and code clean up.
